# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Shift Sensei (シフト先生) - a fast monthly shift-scheduling tool for nursing home unit leaders
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shift-types:
    get:
      operationId: listShiftTypes
      tags: [shift-types]
      summary: List all shift types, ordered by their display/sort order
      responses:
        "200":
          description: List of shift types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftType"
    post:
      operationId: createShiftType
      tags: [shift-types]
      summary: Create a shift type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftTypeInput"
      responses:
        "201":
          description: Created shift type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftType"
  /shift-types/reorder:
    post:
      operationId: reorderShiftTypes
      tags: [shift-types]
      summary: Set the display/sort order of shift types
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftTypeReorderBody"
      responses:
        "200":
          description: Shift types in their new order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftType"
  /shift-types/{id}:
    patch:
      operationId: updateShiftType
      tags: [shift-types]
      summary: Update a shift type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftTypeUpdate"
      responses:
        "200":
          description: Updated shift type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftType"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteShiftType
      tags: [shift-types]
      summary: Delete a shift type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /night-duty-ownership:
    get:
      operationId: listNightDutyOwnership
      tags: [night-duty-ownership]
      summary: List night-duty ownership records, optionally filtered by month
      parameters:
        - name: month
          in: query
          required: false
          description: "Filter by month in YYYY-MM format"
          schema:
            type: string
      responses:
        "200":
          description: List of night-duty ownership records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NightDutyOwnership"
    post:
      operationId: upsertNightDutyOwnership
      tags: [night-duty-ownership]
      summary: Set (or clear) which unit owns a given night, by date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NightDutyOwnershipInput"
      responses:
        "200":
          description: The night-duty ownership record for that date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NightDutyOwnership"
...
  /facility-rules:
    get:
      operationId: listFacilityRules
      tags: [facility-rules]
      summary: List facility rules, optionally filtered by unit
      parameters:
        - name: unitId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of facility rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacilityRule"
    post:
      operationId: createFacilityRule
      tags: [facility-rules]
      summary: Create a facility rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilityRuleInput"
      responses:
        "201":
          description: Created facility rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilityRule"
  /facility-rules/{id}:
    patch:
      operationId: updateFacilityRule
      tags: [facility-rules]
      summary: Update (or toggle enabled on) a facility rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilityRuleUpdate"
      responses:
        "200":
          description: Updated facility rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilityRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFacilityRule
      tags: [facility-rules]
      summary: Delete a facility rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ShiftTypeCategory:
      type: string
      description: >
        Broad bucket a shift type belongs to. Not currently used to enforce
        any scheduling rule (e.g. consecutive-work limits) — it only shapes
        the data so such rules can be added later.
      enum: [day_shift, night_shift, after_night, holiday, paid_leave, other]
    ShiftType:
      type: object
      description: "An editable shift type / stamp definition shown in the schedule grid's toolbar"
      required:
        [
          id,
          code,
          name,
          shortLabel,
          bgColor,
          textColor,
          startTime,
          endTime,
          workingHours,
          category,
          sortOrder,
          countsAsConsecutiveWork,
          canAiAutoAssign,
          includeInMonthlyStats,
          constraintOnly,
          balanceTolerance,
          createdAt,
        ]
      properties:
        id:
          type: integer
        code:
          type: string
          description: "Unique value stored in Shift.code when this type is assigned"
        name:
          type: string
        shortLabel:
          type: string
          description: "1-3 character label rendered on the stamp/grid cell"
        bgColor:
          type: string
          description: "Hex background color for the stamp"
        textColor:
          type: string
          description: "Hex text color for the stamp"
        startTime:
          oneOf:
            - type: string
            - type: "null"
        endTime:
          oneOf:
            - type: string
            - type: "null"
        workingHours:
          description: "Working hours this shift type counts toward, if any"
          oneOf:
            - type: number
            - type: "null"
        category:
          $ref: "#/components/schemas/ShiftTypeCategory"
        sortOrder:
          type: integer
        countsAsConsecutiveWork:
          type: boolean
          description: >
            AI Settings: whether a day stamped with this shift type counts
            toward consecutive-work-day calculations. Used today by Shift
            Sensei's consecutive-work calculations.
        canAiAutoAssign:
          type: boolean
          description: >
            AI Settings: whether an automated scheduler may assign this
            shift type on its own. Not enforced anywhere yet — reserved for
            future automatic scheduling.
        includeInMonthlyStats:
          type: boolean
          description: >
            AI Settings: whether this shift type is counted in monthly
            workload/balance statistics. Not enforced anywhere yet.
        constraintOnly:
          type: boolean
          description: >
            AI Settings: marks a shift type that only ever represents a
            constraint (e.g. a hope-off/must-work marker), never a real
            assignable work shift. Not enforced anywhere yet.
        balanceTolerance:
          type: integer
          description: >
            Acceptable difference (±) from the per-staff average count of
            this shift type in a month, used by the recommendation engine's
            shift-balance check. Edited on the Month-End Check Settings
            page.
        createdAt:
          type: string
          format: date-time
    ShiftTypeInput:
      type: object
      required: [code, name, shortLabel, category]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        shortLabel:
          type: string
          minLength: 1
          maxLength: 3
        bgColor:
          type: string
        textColor:
          type: string
        startTime:
          oneOf:
            - type: string
            - type: "null"
        endTime:
          oneOf:
            - type: string
            - type: "null"
        workingHours:
          oneOf:
            - type: number
            - type: "null"
        category:
          $ref: "#/components/schemas/ShiftTypeCategory"
        sortOrder:
          type: integer
        countsAsConsecutiveWork:
          type: boolean
        canAiAutoAssign:
          type: boolean
        includeInMonthlyStats:
          type: boolean
        constraintOnly:
          type: boolean
        balanceTolerance:
          type: integer
    ShiftTypeUpdate:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        shortLabel:
          type: string
          minLength: 1
          maxLength: 3
        bgColor:
          type: string
        textColor:
          type: string
        startTime:
          oneOf:
            - type: string
            - type: "null"
        endTime:
          oneOf:
            - type: string
            - type: "null"
        workingHours:
          oneOf:
            - type: number
            - type: "null"
        category:
          $ref: "#/components/schemas/ShiftTypeCategory"
        sortOrder:
          type: integer
        countsAsConsecutiveWork:
          type: boolean
        canAiAutoAssign:
          type: boolean
        includeInMonthlyStats:
          type: boolean
        constraintOnly:
          type: boolean
        balanceTolerance:
          type: integer
    ShiftTypeReorderBody:
      type: object
      description: "Full list of shift type ids in their new display order"
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    NightDutyOwnership:
      type: object
      description: >
        Internal only (no dedicated UI yet): records which unit owns a
        given night's shift before an individual staff member is assigned.
      required: [id, date, unitId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        date:
          type: string
          description: "Date in YYYY-MM-DD format"
        unitId:
          oneOf:
            - type: integer
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NightDutyOwnershipInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
        unitId:
          description: "Set to null to clear ownership for that date"
          oneOf:
            - type: integer
            - type: "null"
...
    FacilityRuleType:
      type: string
      description: >
        Selects which evaluator in the facility rule engine interprets
        `config`. "staffing_count" requires a min/max number of staff on a
        shift code that day, optionally gated on another code's
        presence/absence. "staff_day_restriction" and "staff_fixed_shift"
        are produced by "Teach Shift Sensei" from AI相談 conversation —
        new rule types can keep being added later without breaking
        existing rules.
      enum: [staffing_count, staff_day_restriction, staff_fixed_shift]
    FacilityRuleSource:
      type: string
      description: >
        "facility" rules are configured directly on the Facility Rules
        screen. "learned" rules were accepted from a "Teach Shift Sensei"
        suggestion in AI相談.
      enum: [facility, learned]
    FacilityRuleCondition:
      type: object
      description: >
        Gates a "staffing_count" rule to only apply on days where the
        presence/absence of another shift code matches `mode` (e.g. only
        require min 1 "A" on days with no "明け"). Omit entirely for a rule
        that applies every day.
      required: [codes, mode]
      properties:
        codes:
          type: array
          items:
            type: string
          minItems: 1
        mode:
          type: string
          enum: [present, absent]
    FacilityRuleConfig:
      type: object
      description: >
        Which fields are meaningful depends on `ruleType` — see
        FacilityRuleType. All fields are optional so the same object shape
        can carry the config for any rule type.
      properties:
        targetCodes:
          description: "staffing_count: shift codes counted toward satisfying the rule (e.g. [\"A\"], or [\"2夜\", \"特2夜\"] to treat both as one night-shift group)"
          type: array
          items:
            type: string
          minItems: 1
        comparison:
          description: "staffing_count only"
          type: string
          enum: [min, max]
        count:
          description: "staffing_count only"
          type: integer
          minimum: 0
        condition:
          description: "staffing_count only"
          oneOf:
            - $ref: "#/components/schemas/FacilityRuleCondition"
            - type: "null"
        staffId:
          description: "staff_day_restriction / staff_fixed_shift: which staff member this rule is about"
          type: integer
        dayOfWeek:
          description: "staff_day_restriction only: 0 (Sunday) through 6 (Saturday)"
          type: integer
          minimum: 0
          maximum: 6
        disallowedCodes:
          description: "staff_day_restriction only: shift codes this staff member must not receive on dayOfWeek"
          type: array
          items:
            type: string
        allowedCodes:
          description: "staff_fixed_shift only: the only shift codes this staff member should receive (holiday/paid-leave codes are always allowed)"
          type: array
          items:
            type: string
    FacilityRule:
      type: object
      required: [id, unitId, ruleType, source, name, description, enabled, config, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        unitId:
          oneOf:
            - type: integer
            - type: "null"
        ruleType:
          $ref: "#/components/schemas/FacilityRuleType"
        source:
          $ref: "#/components/schemas/FacilityRuleSource"
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        config:
          $ref: "#/components/schemas/FacilityRuleConfig"
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FacilityRuleInput:
      type: object
      required: [name, config]
      properties:
        unitId:
          oneOf:
            - type: integer
            - type: "null"
        ruleType:
          $ref: "#/components/schemas/FacilityRuleType"
        source:
          $ref: "#/components/schemas/FacilityRuleSource"
        name:
          type: string
          minLength: 1
        description:
          type: string
        enabled:
          type: boolean
        config:
          $ref: "#/components/schemas/FacilityRuleConfig"
        sortOrder:
          type: integer
    FacilityRuleUpdate:
      type: object
      description: "All fields optional — used both for full edits and for a bare {enabled} toggle"
      properties:
        unitId:
          oneOf:
            - type: integer
            - type: "null"
        ruleType:
          $ref: "#/components/schemas/FacilityRuleType"
        source:
          $ref: "#/components/schemas/FacilityRuleSource"
        name:
          type: string
          minLength: 1
        description:
          type: string
        enabled:
          type: boolean
        config:
          $ref: "#/components/schemas/FacilityRuleConfig"
        sortOrder:
          type: integer